lots of stuf

+++ server/core/src/model/asset.rs
@@ -65,7 +65,7 @@ pub struct CreateAssetBase {
     pub timestamp_info: TimestampInfo,
     pub size: Size,
     /// degrees clockwise
-    pub rotation_correction: Option<i32>,
+    pub rotation_correction: Option<i64>,
     /// Seahash of the file, if already computed
     pub hash: Option<u64>,
     pub gps_coordinates: Option<GpsCoordinates>,
