+++ server/core/src/model/repository/db_entity/asset.rs
@@ -16,7 +16,7 @@ pub struct DbAsset {
     pub timezone_info: DbTimestampInfo,
     pub width: i64,
     pub height: i64,
-    pub rotation_correction: Option<i32>,
+    pub rotation_correction: Option<i64>,
     pub gps_latitude: Option<i64>,
     pub gps_longitude: Option<i64>,
     pub thumb_small_square_avif: i64,
